TASK-5449 Add the ability to define local variables in argus

+++ dargus/argus.py
@@ -1,7 +1,5 @@
 import os
 import logging
-import random
-import string
 import importlib.util
 
 import yaml
@@ -15,7 +13,7 @@ from dargus.test import Test
 from dargus.step import Step
 from dargus.validator import Validator
 from dargus.validation_result import ValidationResult
-from dargus.utils import get_item_from_json, create_url
+from dargus.utils import get_item_from_json, create_url, replace_random_vars, replace_variables
 from dargus.commons import query
 
 
@@ -23,27 +21,9 @@ LOGGER = logging.getLogger('argus_logger')
 
 
 class Argus:
-    def __init__(self, suite_dir, argus_config, output_prefix=None, output_dir=None):
+    def __init__(self, argus_config):
 
-        # Getting suite directory
-        self.suite_dir = os.path.realpath(os.path.expanduser(suite_dir))
-
-        # Getting argus configuration
         self.config = argus_config
-
-        # Setting up output directory
-        if output_dir is None:
-            self.out_fpath = suite_dir
-        else:
-            out_fpath = os.path.realpath(os.path.expanduser(output_dir))
-            os.makedirs(out_fpath, exist_ok=True)
-            self.out_fpath = out_fpath
-
-        # Setting up output file names
-        if output_prefix is None:
-            self.out_prefix = 'argus_out_' + datetime.now().strftime('%Y%m%d%H%M%S')
-        else:
-            self.out_prefix = output_prefix
 
         self.suites = []
 
@@ -55,7 +35,6 @@ class Argus:
         self.validation_results = []
 
         self._generate_token()
-        self._parse_files(self.suite_dir)
 
         # Loading validator
         if 'validator' in self.config and self.config['validator'] is not None:
@@ -71,6 +50,9 @@ class Argus:
             self.validator = validator_class(config=self.config, auth_token=self.auth_token)
         else:
             self.validator = Validator(config=self.config, auth_token=self.auth_token)
+
+        # Parsing validation files
+        self._parse_files(self.config['suiteDir'])  # files must be parsed after validator logs in
 
     @staticmethod
     def _login(auth, field):
@@ -92,50 +74,83 @@ class Argus:
             else:
                 self.auth_token = authentication['token']
 
-    def _parse_files(self, test_folder):
-        fpaths = [os.path.join(test_folder, file)
-                  for file in os.listdir(test_folder)
-                  if os.path.isfile(os.path.join(test_folder, file)) and
-                  file.endswith('.yml')]
+    def _select_suites_to_run(self, fpaths):
+        selected_suites = []
         for fpath in fpaths:
-            LOGGER.debug('Parsing file "{}"'.format(fpath))
-            with open(fpath, 'r') as fhand:
+            # Getting suite ID (basename without last file extension)
+            suite_id = '.'.join(os.path.basename(fpath).split('.')[:-1])
+
+            # Checking if duplicated ID
+            if suite_id in self.suite_ids:
+                raise ValueError('Duplicated suite ID "{}"'.format(suite_id))
+            self.suite_ids.append(suite_id)
+
+            # Filtering suites to run with regex support
+            if self.config['suites']:
+                for s in self.config['suites']:
+                    match = re.findall('^' + s + '$', suite_id)
+                    if match and match[0] == suite_id:
+                        selected_suites.append({'id': suite_id, 'fpath': fpath})
+            else:
+                selected_suites.append({'id': suite_id, 'fpath': fpath})
+
+        return selected_suites
+
+    def _parse_files(self, suite_dir):
+        # Getting all yml files from suite folder
+        fpaths = [os.path.join(suite_dir, file)
+                  for file in os.listdir(suite_dir)
+                  if os.path.isfile(os.path.join(suite_dir, file)) and
+                  file.endswith('.yml')]
+
+        # Filtering suites to run
+        selected_suites = self._select_suites_to_run(fpaths)
+
+        # Parsing suite file
+        for selected_suite in selected_suites:
+            suite_id, suite_fpath = selected_suite['id'], selected_suite['fpath']
+            LOGGER.debug('Parsing file "{}"'.format(suite_fpath))
+            with open(suite_fpath, 'r') as suite_fhand:
+                # Replacing RANDOM template functions
+                replace_random_vars(suite_fhand.readlines())
+                suite_fhand.seek(0)
+
+                # Loading YAML
                 try:
-                    suite = yaml.safe_load(fhand)
+                    suite_content = yaml.safe_load(suite_fhand)
                 except yaml.parser.ParserError as e:
                     msg = 'Skipping file "{}". Unable to parse YML file. {}.'
-                    LOGGER.error(msg.format(fpath, ' '.join(str(e).replace('\n', ' ').split()).capitalize()))
+                    LOGGER.error(msg.format(suite_fpath, ' '.join(str(e).replace('\n', ' ').split()).capitalize()))
                     continue
-            suite = self._parse_suite(suite)
+
+            suite = self._parse_suite(suite_id, suite_content)
             if suite is not None:
                 self.suites.append(suite)
 
-    def _parse_suite(self, suite):
-        # Getting suite ID
-        id_ = suite.get('id')
-        if id_ is None:
-            raise ValueError('Field "id" is required for each suite')
-        if id_ in self.suite_ids:
-            raise ValueError('Duplicated suite ID "{}"'.format(id_))
-        self.suite_ids.append(id_)
-
-        # Filtering suites to run
-        if 'suites' in self.config and self.config['suites'] is not None:
-            if id_ not in self.config['suites']:
-                return None
-
+    def _parse_suite(self, suite_id, suite):
         # Getting base URL
         if suite.get('baseUrl') is None and 'baseUrl' in self.config:
             suite['baseUrl'] = self.config['baseUrl']
         base_url = suite.get('baseUrl')
 
-        tests = list(filter(None, [self._parse_test(test) for test in suite.get('tests')]))
+        # Getting suite output dir
+        output_dir = suite.get('outputDir') or os.path.join(self.config['outputDir'], suite_id)
 
-        suite = Suite(id_=id_, base_url=base_url, tests=tests)
+        # Getting suite parameters
+        name = suite.get('name') or suite_id
+        description = suite.get('description')
+        suite_variables = suite.get('variables') or {}
+
+        # Getting tests
+        tests = list(filter(None, [self._parse_test(test, suite_variables) for test in suite.get('tests')]))
+
+        # Creating suite
+        suite = Suite(id_=suite_id, name=name, description=description, base_url=base_url, variables=suite_variables,
+                      tests=tests, output_dir=output_dir)
 
         return suite
 
-    def _parse_test(self, test):
+    def _parse_test(self, test, suite_variables):
         # Getting test ID
         id_ = test.get('id')
         if id_ is None:
@@ -144,10 +159,17 @@ class Argus:
             raise ValueError('Duplicated test ID "{}"'.format(id_))
         self.test_ids.append(id_)
 
+        description = test.get('description')
+        test_variables = test.get('variables') or {}
         tags = test.get('tags').split(',') if test.get('tags') else None
         path = test.get('path')
         method = test.get('method')
         async_ = test.get('async')
+
+        # Run specific tags if defined
+        if self.config['tags']:
+            if tags is None or (not set(tags).intersection(set(self.config['tags']))):
+                return None
 
         # Filtering tests to run
         if 'validation' in self.config and self.config['validation'] is not None:
@@ -173,9 +195,10 @@ class Argus:
 
         steps = []
         for step in test.get('steps'):
-            steps += list(filter(None, self._parse_step(step)))
+            steps += list(filter(None, self._parse_step(step, suite_variables, test_variables)))
 
-        test = Test(id_=id_, tags=tags, path=path, method=method, headers=headers, async_=async_, steps=steps)
+        test = Test(id_=id_, description=description, variables=test_variables, tags=tags, path=path, method=method,
+                    headers=headers, async_=async_, steps=steps)
         return test
 
     @staticmethod
@@ -230,27 +253,7 @@ class Argus:
 
         return body_params_list
 
-    @staticmethod
-    def replace_template_vars(params):
-        if params:
-            for param in params:
-                if isinstance(params[param], str) and '${' in params[param]:
-                    template, func, args = re.findall('.*(\${(.*\((.*)\))}).*', params[param])[0]
-                    if func.startswith('RANDOM'):
-                        n = int(args) if args else 6
-                        random_value = ''.join(random.choices(string.ascii_uppercase + string.digits, k=n))
-                    elif func.startswith('RANDINT'):
-                        a, b = map(int, re.sub(re.compile(r'\s+'), '', args).split(','))
-                        random_value = str(random.randint(a, b))
-                    elif func.startswith('RANDCHOICE'):
-                        choices = re.sub(re.compile(r'\s+'), '', args).split(',')
-                        random_value = random.choice(choices)
-                    else:
-                        raise ValueError('Template function "{}" not supported'.format(template))
-                    params[param] = params[param].replace(template, random_value)
-        return params
-
-    def _parse_step(self, step):
+    def _parse_step(self, step, suite_variables, test_variables):
         # Getting step ID
         id_ = step.get('id')
         if id_ is None:
@@ -259,6 +262,8 @@ class Argus:
             raise ValueError('Duplicated step ID "{}"'.format(id_))
         self.step_ids.append(id_)
 
+        description = step.get('description')
+        step_variables = step.get('variables') or {}
         path_params = step.get('pathParams')
         query_params = step.get('queryParams')
         query_matrix_params = step.get('queryMatrixParams')
@@ -266,6 +271,12 @@ class Argus:
         body_matrix_params = step.get('bodyMatrixParams')
         body_file = step.get('bodyFile')
         validation = step.get('validation')
+
+        # Get variables
+        variables = self.config['variables'].copy()
+        variables.update(suite_variables)
+        variables.update(test_variables)
+        variables.update(step_variables)
 
         # Parsing matrix params
         if query_matrix_params is not None:
@@ -283,12 +294,16 @@ class Argus:
                         query_params[key] = default_params[key]
 
         # Parsing body params
+        if variables:
+            body_file = replace_variables(body_file, variables)
         body_params_list = self._parse_body(id_, body_params, body_matrix_params, body_file)
 
-        # Replace template variables
-        path_params = self.replace_template_vars(path_params)
-        query_params_list = [self.replace_template_vars(params) if params else None for params in query_params_list]
-        body_params_list = [self.replace_template_vars(params) if params else None for params in body_params_list]
+        # Replacing variables
+        if variables:
+            path_params = replace_variables(path_params, variables)
+            query_params_list = replace_variables(query_params_list, variables)
+            body_params_list = replace_variables(body_params_list, variables)
+            validation = replace_variables(validation, variables)
 
         # Cartesian product between query and body params
         step_params = [i for i in product(query_params_list, body_params_list)]
@@ -300,24 +315,26 @@ class Argus:
 
         # Creating steps
         steps = [
-            Step(id_=id_, path_params=path_params, query_params=step_params[i][0], body_params=step_params[i][1],
-                 validation=validation)
+            Step(id_=id_, description=description, variables=step_variables, path_params=path_params,
+                 query_params=step_params[i][0], body_params=step_params[i][1], validation=validation)
             for i, id_ in enumerate(id_list)
         ]
 
         return list(filter(None, steps))
 
-    def get_validation_results(self, response, current, url, headers):
+    def _get_validation_results(self, response, current, url, headers):
         # Validating response
         validation = []
         if not current.tests[0].async_:  # Non-asynchronous queries
             response_is_valid, events = self.validator.validate_response(response)
             if response_is_valid:
                 validation = self.validator.validate(response, current)
+            self.validator.run_after_validation(response, current)
         else:  # Asynchronous queries
             response_is_valid, events = self.validator.validate_async_response(response)
             if response_is_valid:
                 validation = self.validator.validate(response, current)
+            self.validator.run_after_async_validation(response, current)
 
         # Creating validation result
         vr = ValidationResult(current=current,
@@ -328,22 +345,27 @@ class Argus:
                               headers=headers)
         self.validation_results.append(vr)
 
-    def write_output(self):
+    def _write_output(self, suite):
         """Write validation results in different file formats"""
 
+        # Setting up timestamp for file names
+        timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
+
         # Writing to JSON file
-        out_fpath_json = os.path.join(self.out_fpath, self.out_prefix + '.json')
+        out_fpath_json = os.path.join(suite.output_dir, '{}_{}.json'.format(suite.id_, timestamp))
         LOGGER.debug('Writing results to "{}"'.format(out_fpath_json))
-        out_fhand = open(out_fpath_json, 'w')
-        out_fhand.write('\n'.join([json.dumps(vr.to_json()) for vr in self.validation_results]) + '\n')
-        out_fhand.close()
+        if not self.config['dry_run']:
+            out_fhand = open(out_fpath_json, 'w')
+            out_fhand.write('\n'.join([json.dumps(vr.to_json()) for vr in self.validation_results]) + '\n')
+            out_fhand.close()
 
         # Writing to HTML file
-        out_fpath_html = os.path.join(self.out_fpath, self.out_prefix + '.html')
+        out_fpath_html = os.path.join(suite.output_dir, '{}_{}.html'.format(suite.id_, timestamp))
         LOGGER.debug('Writing results to "{}"'.format(out_fpath_html))
-        out_fhand = open(out_fpath_html, 'w')
-        out_fhand.write('\n'.join([vr.to_html() for vr in self.validation_results]) + '\n')
-        out_fhand.close()
+        if not self.config['dry_run']:
+            out_fhand = open(out_fpath_html, 'w')
+            out_fhand.write('\n'.join([vr.to_html() for vr in self.validation_results]) + '\n')
+            out_fhand.close()
 
     def execute(self):
         """
@@ -355,6 +377,7 @@ class Argus:
         """
 
         for suite in self.suites:
+            os.makedirs(suite.output_dir, exist_ok=True)  # Setting up output directory for the suite
             current = suite
             for test in suite.tests:
                 current.tests = [test]
@@ -374,6 +397,12 @@ class Argus:
                     # Querying current suite-test-step
                     LOGGER.debug('Querying: Suite "{}"; Test "{}"; Step "{}"'.format(suite.id_, test.id_, step.id_))
                     LOGGER.debug('Query: {} {} {}'.format(method, url, body))
+
+                    # Stop if dry-run
+                    if self.config['dry_run']:
+                        continue
+
+                    # Querying
                     if not current.tests[0].async_:  # Non-asynchronous queries
                         response = query(url=url, method=method, headers=headers, body=body)
                     else:  # Asynchronous queries
@@ -387,7 +416,7 @@ class Argus:
 
                     # Validating results
                     LOGGER.debug('Validating: Suite "{}"; Test "{}"; Step "{}"'.format(suite.id_, test.id_, step.id_))
-                    self.get_validation_results(response, current, url, headers)
+                    self._get_validation_results(response, current, url, headers)
 
-        # Writing output
-        self.write_output()
+            # Writing output
+            self._write_output(suite)
